Refactor Map into Fragment
