+++ app/src/main/java/com/sdpteam/connectout/map/MapFragment.java
@@ -1,0 +1,76 @@
+package com.sdpteam.connectout.map;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.ImageButton;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.SupportMapFragment;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
+import com.sdpteam.connectout.R;
+
+import java.util.List;
+
+public class MapFragment extends Fragment implements OnMapReadyCallback {
+
+    private GoogleMap mMap;
+
+    private MapViewModel mapViewModel;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.activity_map, container, false);
+
+        // Obtain the SupportMapFragment and get notified when the map is ready to be used.
+        //container object for the map used for the map lifecycle management and UI capabilities
+        SupportMapFragment mapFragment = (SupportMapFragment) getChildFragmentManager().findFragmentById(R.id.map);
+        mapFragment.getMapAsync(this);
+
+        mapViewModel = new ViewModelProvider(requireActivity()).get(MapViewModel.class);
+        mapViewModel.init(new MapModel());
+
+        mapViewModel.getEventList().observe(getViewLifecycleOwner(), eventList -> showNewMarkerList(eventList));
+
+        ImageButton refreshButton = rootView.findViewById(R.id.refresh_button);
+        refreshButton.setOnClickListener(view -> showNewMarkerList(mapViewModel.refreshEventList().getValue()));
+
+        return rootView;
+    }
+
+    private void showNewMarkerList(List<Event> eventList) {
+        if (mMap == null) {
+            return;
+        }
+        mMap.clear();
+        for (Event e : eventList) {
+            LatLng ll = new LatLng(e.getLat(), e.getLng());
+            MarkerOptions m = new MarkerOptions().position(ll).title(e.getTitle());
+            mMap.addMarker(m);
+        }
+    }
+
+    /**
+     * Manipulates the map once available.
+     * This callback is triggered when the map is ready to be used.
+     * This is where we can add markers or lines, add listeners or move the camera.
+     * If Google Play services is not installed on the device, the user will be prompted to install
+     * it inside the SupportMapFragment. This method will only be triggered once the user has
+     * installed Google Play services and returned to the app.
+     */
+    @Override
+    public void onMapReady(@NonNull GoogleMap googleMap) {
+        mapViewModel.onMapReady(googleMap);
+        mMap = googleMap;
+    }
+}
